# Remove most string allocations when reading file line by line