--- a/include/chemfiles/files/PlainFile.hpp
+++ b/include/chemfiles/files/PlainFile.hpp
@@ -20,9 +20,9 @@ public:
     ~PlainFile() override;
 
     size_t read(char* data, size_t count) override;
-    size_t write(const char* data, size_t count) override;
+    void write(const char* data, size_t count) override;
 
-    void clear() override;
+    void clear() noexcept override;
     void seek(int64_t position) override;
 
 private: