--- a/include/chemfiles/files/PlainFile.hpp
+++ b/include/chemfiles/files/PlainFile.hpp
@@ -23,7 +23,7 @@ public:
     void write(const char* data, size_t count) override;
 
     void clear() noexcept override;
-    void seek(int64_t position) override;
+    void seek(uint64_t position) override;
 
 private:
     std::FILE* file_;