… .codecov.yml
… CMakeLists.txt
… external/molfiles.tar.gz
… external/zlib.tar.gz
… include/chemfiles/error_fmt.hpp
… include/chemfiles/formats/GRO.hpp
… include/chemfiles/formats/LAMMPSData.hpp
… include/chemfiles/formats/MMTF.hpp
… include/chemfiles/formats/MOL2.hpp
… include/chemfiles/formats/Molfile.hpp
… include/chemfiles/formats/PDB.hpp
… include/chemfiles/formats/SDF.hpp
… include/chemfiles/formats/SMI.hpp
… include/chemfiles/formats/TNG.hpp
… include/chemfiles/formats/TRR.hpp
… include/chemfiles/formats/Tinker.hpp
… include/chemfiles/formats/XTC.hpp
… include/chemfiles/formats/XYZ.hpp
… include/chemfiles/formats/mmCIF.hpp
… include/chemfiles/parse.hpp
… include/chemfiles/selections/expr.hpp
… include/chemfiles/string_view.hpp
… include/chemfiles/utils.hpp
… src/Configuration.cpp
… src/Connectivity.cpp
… src/File.cpp
… src/Format.cpp
… src/FormatFactory.cpp
… src/Frame.cpp
… src/Property.cpp