@Luthaf 2019-10-14
Finally manage to make all compilers happy \o/