@frodofine 2019-10-18
This is a large PR, I'll try to get to it when I can.