@Luthaf 2019-11-07
I added a few improvements if you want to look at them @ezavod. Else I'll merge this tomorrow =)